[v1.0.9](https://github.com/DKWechatHelper/DKWechatHelper/releases/tag/1.0.9) / 2021-09-15 what's new * 好友检测改为通过请求是否可以转账进行判断 * 动态启动图资源改为下载 * 支持8.0.13 * 更新越狱包8.0.13 * 更新已注入助手的8.0.13未签名包 * 更新越狱源安装包

+++ dkhelper/dkhelperDylib/DKCleanFriendsController.m
@@ -131,7 +131,7 @@
         });
     }else{
         [DKHelper showAlertWithTitle:@"删除成功"
-                             message:[NSString stringWithFormat: @"已成功删除%d名好友",contactList.count]
+                             message:[NSString stringWithFormat: @"已成功删除%d名好友",(int)contactList.count]
                             btnTitle:@"确定" handler:^(UIButton *sender) { }];
     }
 }
